Ability to select exercise and reps via dropdown
