+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "landingboiler",
-  "version": "0.1.0",
+  "name": "GSquaredFitness",
+  "version": "1.1.0",
   "private": true,
   "dependencies": {
     "firebase": "^6.2.4",
