+++ src/forms/WorkoutForm.js
@@ -1,22 +1,66 @@
 import React, { useState, useContext } from 'react';
-import { ExerciseForm } from "./ExerciseForm"
-import { WorkoutTable, TableTitle } from "../components/WorkoutTable"
-import { PrimaryButton } from '../components/Buttons';
+import { DefaultForm } from "../components/DefaultForm"
 import { toast } from 'react-toastify';
 import { FirebaseContext } from '../config/Firebase/FirebaseContext';
 
 export const WorkoutForm = ({ setSelection, uid }) => {
-
     const firebase = useContext(FirebaseContext)
-    const [exerciseList, addExercise] = useState([])
+    const defaultState = [
+        {
+            name: 'exercise',
+            value: 'Select',
+            label: 'Select Exercise',
+            options: ['Squat', 'Bench', 'Overhead Press', 'Bent Over Row', 'Deadlift'],
+            free: false,
+
+        },
+        {
+            name: 'weight',
+            value: '',
+            placeholder: 'Weight Used (lbs)',
+            type: 'number',
+            free: true
+        },
+        {
+            name: 'reps',
+            value: 'Select',
+            label: 'Select Reps',
+            options: [1, 3, 5, 10],
+            free: false
+        }
+    ]
+
+
     const [errors, setErrors] = useState({})
+    const [values, setValues] = useState(defaultState)
 
-    async function onSubmit() {
+    function onChange(event) {
+        const { value } = event.target;
+        const changeIndex = values.findIndex(({ name }) => name === event.target.name)
+        const changedObject = { ...values[changeIndex], value: value }
+        setValues(preValues => (
+            [...preValues.slice(0, changeIndex),
+                changedObject,
+            ...preValues.slice(changeIndex + 1)
+            ]
+        ));
+
+    }
+
+    async function onSubmit(event) {
+
+        event.preventDefault();
+        if (!formIsValid()) return;
+
+        const workout = {}
+        workout.exercise = values[0].value;
+        workout.weight = values[1].value;
+        workout.reps = values[2].value;
         const timestamp = JSON.stringify(Date.now())
-
         try {
-            await firebase.db.ref('/workouts/' + uid).push({
-                exerciseList,
+            setValues(defaultState)
+            await firebase.db.ref('/maxes/' + uid).push({
+                workout,
                 timestamp
             })
             setSelection(0)
@@ -30,6 +74,19 @@ export const WorkoutForm = ({ setSelection, uid }) => {
         }
     }
 
+
+    function formIsValid() {
+        const exercise = values[0].value;
+        const weight = values[1].value;
+        const reps = values[2].value;
+        const errors = {};
+        if (!weight) errors.weight = "Weight is required.";
+        if (!exercise) errors.exercise = "Exercise is required.";
+        if (!reps) errors.reps = "Rep number is required.";
+        setErrors(errors);
+        return Object.keys(errors).length === 0;
+    }
+
     return (
         <div style={{
             display: 'flex',
@@ -37,23 +94,17 @@ export const WorkoutForm = ({ setSelection, uid }) => {
             alignItems: 'center',
             justifyContent: 'center',
         }}>
-            {exerciseList.length > 0 ?
-                <>
-                    <TableTitle>Workout</TableTitle>
-                    <WorkoutTable
-                        tableContent={exerciseList} />
-                    <PrimaryButton onClick={() => onSubmit()}> Submit Workout </PrimaryButton>
-                    <br />
-                    <br />
-                </> : <p style={{ marginTop: "0px" }}>Add an exercise below.</p>
-            }
-            <ExerciseForm
-                exerciseList={exerciseList}
-                addExercise={addExercise}
+            <DefaultForm
+                onChange={onChange}
+                onSubmit={onSubmit}
+                values={values}
+                errors={errors}
+                formHeading='Add Workout'
+                buttonText='Submit Workout'
             />
-
-
         </div>
-
     )
 }
+
+
+
