+++ src/pages/Dashboard.js
@@ -1,7 +1,5 @@
 import React, { useContext, useState } from 'react';
 import { AuthContext } from '../session/AuthContext';
-import RunningReport from "../reports/RunningReport"
-import WeightReport from "../reports/WeightReport"
 import WorkoutReport from "../reports/WorkoutReport"
 import { FormSelector } from "../components/FormSelector"
 import report from "../assets/images/report.svg"
@@ -10,10 +8,8 @@ const Dashboard = () => {
     const [selection, setSelection] = useState(0);
     const { authState: { user } } = useContext(AuthContext);
     const { uid } = user || JSON.parse(localStorage.getItem("user"))
-    const formOptions = ['Runs', 'Weight', 'Workouts']
+    const formOptions = ['Workouts']
     const reports = [
-        <RunningReport setSelection={setSelection} uid={uid} />,
-        <WeightReport setSelection={setSelection} uid={uid} />,
         <WorkoutReport setSelection={setSelection} uid={uid} />
     ]
